+++ CHANGELOG.md
@@ -3,11 +3,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### [0.8.6] - 2020-11-28
+
+#### Changed
+
+* Props now called with (component) if it is a function.
+* Event args called with (wrapper, event, props, component)
+
 ### [0.8.5] - 2020-11-27
 
-#### Added
+#### Changed
 
-* Callable watched fields now called with (props, view) 
+* Callable watched fields now called with (props, component) 
 
 ### [0.8.4] - 2020-11-24